backend/requirements.txt
django
djangorestframework
markdown
django-cors-headers
django-filter
pyyaml